Fix logic for setting final_Powers[FP_TEAM_HEAL] to 0
Reversed the condition for setting final_Powers[FP_TEAM_HEAL] to 0.
Previously, it was set to 0 if final_Side was FORCE_LIGHTSIDE or
if final_Side was FORCE_DARKSIDE and fpDisabled had the FP_TEAM_HEAL
bit set. Now, it is set to 0 if final_Side is FORCE_DARKSIDE or if
final_Side is FORCE_LIGHTSIDE and fpDisabled has the FP_TEAM_HEAL
bit set.